sharepoint-governance-platform/backend/requirements.txt
# FastAPI Core Dependencies
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Microsoft Integration
msgraph-core==1.0.0
azure-identity==1.15.0
Office365-REST-Python-Client==2.5.3
msal==1.26.0

# Active Directory / LDAP
python-ldap==3.4.4
ldap3==2.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background Jobs
apscheduler==3.10.4
celery[redis]==5.3.6  # Alternative to APScheduler

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Machine Learning (Phase 3)
scikit-learn==1.3.2
joblib==1.3.2

# Reporting
reportlab==4.0.8
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
faker==20.0.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5

# Phase 2: Enhanced Features
reportlab==4.0.7  # PDF generation
openpyxl==3.1.2  # Excel generation
Pillow==10.1.0  # Image processing for reports

# Phase 3: Advanced Features - AI/ML
scikit-learn==1.3.2  # Machine learning
numpy==1.26.2  # Numerical computing
pandas==2.1.3  # Data analysis12.1
mypy==1.8.0

# Security
bandit==1.7.6
safety==3.0.1
